Refactor the rentals logic to a separate directory
