+++ people/main.rb
@@ -15,6 +15,10 @@ class PersonInitialize
       puts ''
     end
   
+    def people 
+      @people
+    end
+    
     def create_person
       print 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
       option = gets.chomp
